Backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
supabase==2.0.0
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2
chardet==5.2.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.24.1
